Reduce data folder cleaner process by skip node that no place file

+++ src/main/java/tannyjung/tanshugetrees_handcode/systems/world_gen/DataFolderCleaner.java
@@ -17,46 +17,57 @@ public class DataFolderCleaner {
         String region = (chunk_pos.x >> 5) + "," + (chunk_pos.z >> 5);
         String node = GameUtils.outside.quardtreeChunkToNode(chunk_pos.x, chunk_pos.z);
         String file_path = dimension + "/" + region + "/" + node + ".txt";
+        File file_place = new File(Handcode.directory_world_data + "/world_gen/place/" + file_path);
 
-        boolean full_node = false;
-        File file_region = new File(Handcode.directory_world_data + "/world_gen/regions/" + file_path);
-        FileManager.writeTXT(file_region.toPath().toString(), "A", true);
+        if (file_place.exists() == true && file_place.isDirectory() == false) {
 
-        // Read to get text length
-        {
+            File file_region = new File(Handcode.directory_world_data + "/world_gen/regions/" + file_path);
+            FileManager.writeTXT(file_region.toPath().toString(), "A", true);
+            boolean full_node = false;
 
-            try { BufferedReader buffered_reader = new BufferedReader(new FileReader(file_region)); String read_all = ""; while ((read_all = buffered_reader.readLine()) != null) {
+            // Read to get text length
+            {
 
-                {
+                try {
+                    BufferedReader buffered_reader = new BufferedReader(new FileReader(file_region));
+                    String read_all = "";
+                    while ((read_all = buffered_reader.readLine()) != null) {
 
-                    if (read_all.length() >= Math.pow(32 >> 2, 2)) {
+                        {
 
-                        full_node = true;
+                            if (read_all.length() >= (32 >> 2) * (32 >> 2)) {
+
+                                full_node = true;
+
+                            }
+
+                        }
 
                     }
-
+                    buffered_reader.close();
+                } catch (Exception exception) {
+                    OutsideUtils.exception(new Exception(), exception);
                 }
 
-            } buffered_reader.close(); } catch (Exception exception) { OutsideUtils.exception(new Exception(), exception); }
+            }
 
-        }
+            if (full_node == true) {
 
-        if (full_node == true) {
+                File file_detailed_detection = new File(Handcode.directory_world_data + "/world_gen/detailed_detection/" + file_path);
 
-            file_region.delete();
+                file_region.delete();
+                file_place.delete();
+                file_detailed_detection.delete();
 
-            File file_place = new File(Handcode.directory_world_data + "/world_gen/place/" + file_path);
-            File file_detailed_detection = new File(Handcode.directory_world_data + "/world_gen/detailed_detection/" + file_path);
-            file_place.delete();
-            file_detailed_detection.delete();
+                // TanshugetreesMod.LOGGER.debug("Debugging : Delete Place File -> " + chunk_pos.x + " " + chunk_pos.z);
 
-            // TanshugetreesMod.LOGGER.debug("Debugging : Delete Place File -> " + chunk_pos.x + " " + chunk_pos.z);
+                // If finish placing all chunks in region
+                if (file_place.getParentFile().listFiles().length == 0) {
 
-            // If finish placing all chunks in region
-            if (file_place.getParentFile().listFiles() == null || file_place.getParentFile().listFiles().length == 0) {
+                    file_place.getParentFile().delete();
+                    file_detailed_detection.getParentFile().delete();
 
-                file_place.getParentFile().delete();
-                file_detailed_detection.getParentFile().delete();
+                }
 
             }
